fix(quantum-graph): treat nb_obs as a float measure, not an int count
The numba obs-energy kernel built its per-node weights as int32. nb_obs
is a measure weight, not a count: node samplers happen to set integer
counts, but the paper experiments set the fractional population measure
nu (every weight < 1). Casting those to int truncated them all to zero,
so total_obs was 0 and the obs-energy silently collapsed to 0.0 for
every configuration of centers -- a valid-looking number, so nothing
raised.

The damage went beyond the reported energy: MinimizeEnergy selects the
memory-window minimum by comparing sa.calculate_energy with a strict
'<'. With that energy pinned at 0.0, 0.0 < 0.0 was never true, so the
robustification never replaced the initial KMeans++ centers with the
annealed ones. On the grid and the sphere graph route (both driven by
the population measure) the returned 'annealing' result was in fact the
KMeans++ initialization, best-of-N selected downstream by the correct
float energy (node_energy) -- which masked the bug. SBM (integer
nb_obs) and the sphere manifold route (float energy) were unaffected.

The kernel now carries nb_obs as float64. Regression test: a graph with
a fractional measure has a positive obs-energy matching the pure-Python
path exactly.

+++ src/kmeanssa_ng/quantum_graph/space.py
@@ -203,7 +203,7 @@ def _calculate_energy_obs_numba(
     n_points = len(point_positions)
     k = len(center_positions)
     total_energy = 0.0
-    total_obs = 0
+    total_obs = 0.0
 
     for p_idx in range(n_points):
         nb_obs = point_nb_obs[p_idx]
@@ -864,9 +864,14 @@ class QuantumGraph(nx.Graph, Space):
                 self._pairwise_nodes_distance_array,
             )
         else:  # how == "obs"
+            # nb_obs is a measure weight, not an integer count: node samplers
+            # set integer counts, but the paper experiments set the fractional
+            # population measure nu. Casting to int here truncated every
+            # fractional weight to zero, so the whole obs-energy collapsed to
+            # 0.0 -- silently, since 0 is a valid energy. Keep it float.
             point_nb_obs = np.array(
                 [data.get("nb_obs", 0) for _, data in self.nodes(data=True)],
-                dtype=np.int32,
+                dtype=np.float64,
             )
             return _calculate_energy_obs_numba(
                 center_edges_0,
